test: move typechecks to their own test file
